refactor: add query container for user events
